Initial reworking of main.c's code into librvth/rvth.c.
This provides a C interface for opening an RVT-H disk image and
retrieving bank information. rvth.c provides a higher-level abstraction
than the raw disk tables (now referred to as NHCD), which makes it
easier to e.g. print disc names and timestamps.

rvth_open() parses the RVT-H bank table and disc headers in order to get
all of the necessary information. It also checks for deleted disc images
and marks them as such.

Deleted images are identified and marked as such. Note that deleted
dual-layer Wii images are currently identified as single-layer, since
we aren't parsing the FST to determine the highest disc address.