rvth.c: New function rvth_extract().
This function extracts the specified bank to the specified filename.
A callback may be specified to monitor progress.

main.c: Added a third command line parameter for "bank to extract".

If a third parameter, a bank number (1-8), is specified, that bank will
be extracted to BankX.gcm, where X is the bank number.

TODO: Use getopt() or popt for command line parsing.