Implemented the first WRITE command: 'delete'
[librvth] rvth.c: New function rvth_delete(). This function first calls
rvth_make_writable(), which calls ref_make_writable(). Note that the
rvth_make_writable() function first calls ref_is_device() to check if
it's a device file. If it isn't, it fails before even trying to make
it writable, since we don't want to write to an RVT-H disk image.

Tested on an RVT-H Reader, and this command does successfully delete
the specified bank.

Banks that cannot be deleted:
- Empty
- Unknown
- Already deleted banks
- Second bank of a dual-layer Wii disc image