[librvth] rvth.c: Align RVTH reads and writes to sector boundaries.
Windows only allows reading and writing raw devices on sector boundaries,
and we have to read/write the entire sector.

FIXME: We're assuming the partition table for volume group 0 is stored
immediately after the volume group table. This is usually the case, but
we should ensure that it works if this isn't the case.