[librvth] Get the region code for GameCube and Wii games.
New function rvth_init_BankEntry_region() that gets the region code.
Note that it's stored in a different location depending on system.

RvtH_BankEntry: Added a region_code field. This field is 8-bit, though
the original field is 32-bit. This value should never be larger than
0xFF, and if it is, that's an error.

Changed all variables used for reader_read() return values from size_t
to uint32_t, since the function returns the number of LBAs read, not the
number of bytes.

[rvthtool] list-banks.c: Print the region code.

Note that region code 3, "Region-Free", is used by some SDK discs as well
as GCN Action Replay. It is not known to be used by any retail games.