[librvth] Initial attempt to support HDDs with more than 8 banks.
Use the `NHCD_BankTable_Header.bank_count` field to initialize the
`entries` array. We're currently restricting it to a minimum of
8 (default) and a maximum of 32. 32 has not been tested yet.
(I've only tested 9.)

RvtH now has a new field, `is_hdd`, which indicates if the image is an
HDD image or a standalone disc image. Accessor function: rvth_is_hdd()

NHCD_BANK_1_START_LBA() is now a macro, and it takes `bank_count` as a
parameter in order to calculate the default LBA of the first bank.

RvtH_Errors: New error code RVTH_ERROR_INVALID_BANK_COUNT. This error is
returned if the `NHCD_BankTable_Header.bank_count` field is invalid.