[librvth] cert.c: Initial certificate verification code.
cert_verify() takes a ticket or TMD (or anything that starts with a
compatible RSA-2048 signature), decrypts the signature, then verifies
the SHA-1 hash. The SHA-1 hash is calculated based on the rest of the
specified data structure, including the signature issuer field.

Both the signature and SHA-1 hash are checked for validity. If the
calculated hash starts with 0x00 but the signature is invalid, this
usually means it's fakesigned.

On Linux, the certification verification code uses gmp to decrypt the
RSA signature and nettle to calculate the SHA-1 hash. This is currently
not implemented for Windows, so on Windows it will return -ENOSYS.

Note that the Linux build now requires gmp and nettle.

TODO: Verify that it works with nettle-2.x.