[librvth/tests] CertVerifyTest: Verify the debug ticket and TMD certificates.
The "retail" magic appears in these certificates, so remove the check.
On the other hand, the "debug" magic shouldn't appear in retail titles,
so the check has been moved there and changed for debug certificates.

FIXME: The CA certificates are failing with SIG_ERROR_UNSUPPORTED_SIGNATURE_TYPE
and have been disabled in this commit.