[librvth] hashw.h: Added a hash object/context mechanism.
Fakesigning and converting from unencrypted to encrypted requires a lot
of hashing. Creating a hash object once and reusing it allows us to get
rid of some of the overhead involved in the SHA-1 code.

hashw_sha1() is now a wrapper function which can still be used for
simple SHA-1 calculations, e.g. determining if a ticket or TMD is
fakesigned.