[librvth] Refactored error handling in several functions.
Instead of having all of the free() and error handling code duplicated
everywhere, set the error codes and use 'goto' to jump to the end of
the function. This is a common pattern used in the Linux kernel and
many other C programs.

rvth_copy_to_gcm(): Don't store reader_write()'s return value in ret.
This was used for debugging before, and didn't cause a problem until
this refactoring, which uses ret as the actual return value.
- TODO: Check reader_write()'s return value correctly.