Added a precompiled build of mpir-3.0.0 for Windows. (32-bit and 64-bit)
CryptoAPI's RSA functionality is ridiculously complicated to use. Instead
of wasting time figuring out how to use it, we'll just precompile MPIR
(a mostly GMP-compatible library) and nettle for Windows.

This commit includes a precompiled build of mpir-3.0.0. rsaw_gmp.c works
with it with no problems. The Windows version is now significantly faster
at verifying signatures as a result.

Removed rsaw_generic.c, since it's no longer in use.