[librvth] cert.c: New functions to fakesign tickets and TMDs.
These functions zero out the signature area and brute-force bytes in
otherwise unused areas of the relevant structure in order to get an
SHA-1 hash that starts with a 0x00 byte.

NOTE: These functions use different offsets compared to `wit`:
- Ticket: cert.c == 0x25C, `wit` == 0x24C
- TMD:    cert.c == 0x19C, `wit` == 0x19A

cert.c also uses host-endian for the brute-force value, which is the
same as `wit`.

Tested on Dolphin and USB Loader GX on retail hardware.