[librvth] priv_key_store.c: Debug private keys.
These keys consist of the p, q, a, b, and c constants. Technically, only
p and q are needed, since the others can be calculated, but it's easier
to store all five, because we can just import those into nettle directly
without any extra calculations.

rsaw.h: New function rsaw_sha1_sign(). This function takes an RSA-2048
private key and an SHA-1 hash and creates a signature.

TODO: Test this by recrypting a retail Wii game as debug and installing
it on an RVT-H system.