[librvth] rvth_copy_to_hdd(): Set the bank entry's lba_length field.
Otherwise, if we're importing a Wii image over a deleted GameCube image
or vice-versa, we end up with the wrong length. This isn't necessarily
an issue for GCN over Wii, but for Wii over GCN, it can cause the
Wii apploader to crash with a fatal error.

TODO: Importing "The Last Story" into bank 12 caused bank 13 to get
overwritten. Figure out why.