[librvth] rvth_import(): Write the special identifier even if not recrypting.
Also for GameCube games, which would previously cause an error
RVTH_ERROR_NOT_WII_IMAGE to appear due to recryption not being
supported for GameCube.

rvth_create_id(): Parse the Wii partition tables and add the special
identifier to all partitions, including update partitions. Note that
this function is used if importing an RVT-R or unencrypted RVT-H image
to an RVT-H system, which doesn't need recryption.