[librvth] disc_header.c: New file to read GCN/Wii disc headers.
Two new functions:
- rvth_disc_header_identify(): Identify a disc type from an
  already-loaded header.
- rvth_disc_header_get(): Get the header from a file at the specified
  LBA starting address.

rvth_disc_header_identify() returns RVTH_BankType_Unknown even if the
header is empty, so the caller will need to check for emptiness
afterwards.

Tested with an RVT-H HDD image and an RVT-R disc image.

TODO: rvth_disc_header_get(): If the header is empty, check the Wii
partition table to see if we can reconstruct the header from the
game partition. This will either be plaintext for RVT-H images or
encrypted for RVT-R and retail.