[librvth] Initial support for reading standalone disc images in CISO format.
reader_ciso.c: Initial CISO reader. It seems to work for the header for
"The Last Story" (RVT-R), but it isn't reading the rest of the image,
so it can't detect the encryption type. There's probably a bug in the
seeking code.

reader.c: Check for CISO headers, and if found, create a CISO reader.

rvth.c, rvth_open_gcm(): Create the Reader first, then read from that.