[librvth] Re-enabled recryption and fixed several bugs.
rvth_copy_to_gcm_doCrypt():
- Copy both `crypto_type` and `ios_version` when initializing the
  destination bank entry. (Also copy `ios_version` in other places.)
- Initialize the callback state after adjusting lba_copy_len to ensure
  the last line prints correctly.

rvth_extract():
- Check the source crypto type, not the destination crypto type. After
  calling rvth_copy_to_gcm_doCrypt(), the destination data will be
  encrypted using debug keys, but the TMD hasn't been re-signed.
  Checking the destination crypto type would cause the TMD re-sign to
  be skipped, since it's set to Debug. Checking the source crypto type
  works because that's set to None.

rvth_recrypt.c:
- Ticket and TMD are always re-signed regardless of whether or not
  they're recrypted.

The issue I was having with "The Last Story" in Dolphin was actually
caused by an apploader error:

APPLOADER ERROR >>> [production mode] One of the sections in the dol file
exceeded its boundary. All the sections should not exceed 0x80900000
(production mode)

I saw this on the RVT-H Reader, but on devkits, it's handled as a
warning, and the program continues:

APPLOADER WARNING >>> One of the sections in the dol file exceeded its
boundary. All the sections should not exceed 0x80900000 in production
mode.

@JosJuice added a patch to Dolphin (which hasn't been merged yet) to set
the console type to NDEV 2.1 if the disc uses debug keys, which fixes
both the unencrypted and recrypted debug versions of "The Last Story".
(He fixed unencrypted disc reads earlier.)

TODO: Check main.dol file for any issues that would cause an apploader
warning or error.