[librvth] bank_init.c: Read 2 sectors for the DOL header.
Animal Crossing (GAFE01, Rev.00) has the DOL located at 0x1DF84. This is
close to the end of an LBA, so we need to read two LBAs in order to
ensure that we read the entire header. Otherwise, we end up getting
spurious AppLoader errors due to reading garbage data as DOL headers.