[librvth] New enumeration RvtH_ImageType_e.
This enumeration will be used to indicate if the RvtH object is one
of the following:

- RVT-H Reader
- HDD image
- GCM image
- GCM image with SDK headers

This will be used by qrvthtool for more comprehensive icon display.
For example, a GameCube GCM image would be shown with an indigo
GameCube, but the same image with SDK headers would be shown with a
green/cyan GameCube (NR Reader).

This field replaces rvth->is_hdd. rvth_is_hdd() has been updated to
return true for RVT-H Reader and HDD images. All users of rvth->is_hdd
have been updated to either use rvth_is_hdd() or to set the `type` field
with the correct value.