[qrvthtool] Display an icon for GameCube disc images.
An indigo GameCube is used for standalone retail GCMs.

For GCMs with SDK headers, and GameCube games found on RVT-H Reader
HDDs, a green GameCube (NR Reader) is displayed instead.

The GameCube image is from Wikipedia:
https://en.wikipedia.org/wiki/File:GameCube-Bare.jpg

The NR Reader image is a modified version of the GameCube image.
GIMP 2.10.2's Hue-Chroma tool was used to adjust the hue by -110.
This very closely matches online photos of the NR Reader system.

[librvth] Added a function rvth_get_ImageType() to get the RVT-H
image type. This is used to determine which icon is used.