[librvth] ref_get_size(): New function to get a file's size.
Windows apparently doesn't like SEEK_END on devices (or at least a
UCRT update broke it). Use DeviceIoControl() instead.

TODO: Add a flag indicating if a file was deleted in bank table only
or if the disc header was wiped. Also make it possible to read from
an RVT-H Reader with an empty bank table. (assume defaults)