[librvth] If the NHCD table is missing, fake it.
Some RVT-H Readers have a missing NHCD table for some reason, but the
game data is still there. Create a fake NHCD table in memory so we can
extract the banks.

[rvthtool] list-banks.c: Indicate if the NHCD table is missing.

TODO: Add a function to write an empty NHCD table back to the system.
This will be a **DANGEROUS** operation, but it will allow these systems
to be used as an RVT-H Reader again.