[qrvthtool] QRvtHToolWindow: Set the application icon.
The icon is set to RVT-H Reader if either no image is loaded or an
RVT-H Reader device or disk image is loaded. If a standalone image
file is loaded, the icon is set to that image file's icon.

RvtHModel: Reworked the icon functions to make them static. The private
class's QIcon array is now also static. Added wrapper functions to the
public interface and moved the IconID enum to the public class.

Added 16x16, 22x22, and 24x24 icons, which will be used for the window
icon.