[librvth] Rewrote RefFile as a C++ class.
This is the first step in porting librvth to C++, which will make it
easier to maintain. For example, it will allow us to add Apploader
DOL verification for encrypted images more easily, since we will be
able to layer a decrypted partition reader, similar to rom-properties.

Renamed many files to .cpp in order to compile as C++.

TODO:
- RefFile's wrapper functions still use errno instead of m_lastError.
- Remove the C compatibility hack later. (`struct RefFile;`)