[qrvthtool] QRvtHToolWindow: Create the QThread on demand.
It seems that reusing the QThread caused it to invoke the previous worker
object, even though it was deleted. Then again, it was actually deleted
using QObject::deleteLater(), so maybe it wasn't really deleted...

This ensures that even if the old QThread or worker object is still around
for a bit due to QObject::deleteLater(), it won't be invoked twice.