[qrvthtool] Copied over TaskbarButtonManager from mcrecover.
It doesn't seem to work on KDE5. Apparently, the standard task manager
in Plasma 5 uses Unity's protocol, but we don't have libunity on all
systems.

Unity has a D-Bus protocol, but it's supposedly not guaranteed to be
stable. In practice, it is.

QtCreator added support for the D-Bus protocol a while back:
https://codereview.qt-project.org/#/c/33051/2/src/plugins/coreplugin/progressmanager/progressmanager_x11.cpp