[qrvthtool] QRvtHToolWindow::lstBankList_selectionModel_selectionChanged(): Use the `selected` parameter.
The previous method mostly worked, but if the user held down the mouse
button and dragged the selection cursor across the banks, the selected
bank would be the *previously* selected bank, causing the BankEntryView
to get out of sync.

Using the `selected` parameter also eliminates the need to separately
obtain the selection model.