[librvth] deleteBank(), undeleteBank(): Update the timestamp in the in-memory bank entry.
This is needed for qrvthtool in order to show the new timestamp.

writeBankEntry(): Optional parameter `pTimestamp` that returns the
written timestamp in UNIX format. Note that we can't simply use
time(nullptr) here due to timezone conversions, so we end up converting
to NHCD format, then back to UNIX format.