[rvthtool] Added an option to change the IOS version when importing a disc image.
This is useful for e.g. Super Smash Bros. Brawl, which uses IOS36.
The debug version of IOS36 seems to be MIA, so we can change it to
IOS56 and it basically works, though Apploader will show a warning.

TODO: Detect the IOS mismatch and show an Apploader warning here
when listing bank information.