[wadresign] resign-wad.c: Fix copying WAD contents that isn't a multiple of 16 bytes.
While the actual contents might not be a multiple of 16 bytes, AES
operates in blocks of 16 bytes, so we have to copy the extra unused
bytes in order to decrypt the data properly.

This fixes recryption of various IOSes (including IOS58-64-v6176.wad)
and possibly other WADs.

This fixes issue #7: wadresign: Re-signing an IOS misses the last 8
data bytes, resulting in a broken WAD.