Added support for the vWii common key.
It's technically supported in both wadresign and rvthtool, but rvthtool
doesn't allow using the vWii common key for encryption because it doesn't
make sense to recrypt a disc image to vWii.

FIXME: Is there a debug vWii common key?

TODO: A lot of vWii WADs are incorrectly encrypted using the retail
common key, even though the ticket's key index is set correctly.
Add an option to fix the encryption.