[librvth] verify: Add error messages to the callback.
Error messages include hash verification and table copies.

NOTE: We're also checking for zeroed sectors (scrubbed/truncated).
Since this would be zeroed *before* decryption, we have to maintain
two group buffers, one for the encrypted contents, and one for the
decrypted contents. This means 4 MB RAM. Most systems should be able
to handle this in 2022...