verify: Check the partition's data length to determine how much to check in the last group.
The data in the last group past the partition data length has invalid
hashed data. It might be algorithmically generated; I should check
how nkit handles this. For now, ignore it.

Note that we're no longer checking for the "last block group" in the
H3 hash table.

This fixes Sonic Colours (RVT-R), but not swupdate_SDK-3_2_us.gcm.