[nusresign] main.c: New function is_directory().
Split this out of main(). Previously, we were using stat() on all
platforms, but stat() is ANSI on Windows, and it returns 32-bit filesizes.
We only want to use 64-bit filesizes, so we'd have to use _tstati64(),
but we don't actually need any of that; we just need the attributes.

The is_directory() function calls GetFileAttributes() on Windows and
stat() on Linux.