[wadresign] print-info.c: Print an error if a content could not be read.
It seems that some WADs have an incorrect data size value in the header.
This size isn't checked, so re-signing the WAD results in a corrupt
output WAD.

Detect the corrupt WAD when verifying.

NOTE: The exit code was already set to 1 in this case, but no one checks
that when running the command directly instead of through a script.