[qrvthtool] SelectDeviceDialog: Hacky method for auto-refresh on Windows.
Listen for device notifications. If we get a notification that a disk
device was added or removed, refresh the entire list. This isn't as
nice as the udev method, since that one only adds or removes the
specific device, but it works for now.

TODO: Make the Windows one a listener by creating a hidden window?