[librvth] extract.cpp: Call flush() after every write().
This is basically the same as O_SYNC, though we have up to a 1 MB buffer
that gets written to disk at once.

I tried with 32 MB flushing at first, but it just resulted in the
progress bar going up in 32 MB increments with freezes due to the
32 MB hitting the I/O cache first.

RefFile::flush(): Call fsync() on Linux and FlushFileBuffers() on Windows
to ensure the file is actually flushed to disk. fflush() only clears the
stdio buffers.

TODO: Use Win32 and/or POSIX I/O directly? I tried this with open(),
lseek(), read(), write(), etc., but it wasn't able to read the ticket/TMD
signatures for some reason. This solution should be good enough for now.