added show-table command, for inspecting NHCDTable
since you mentioned wanting to add in a command for repairing a
NHCD Bank Table, I figured a command to inspect your current NHCD
Bank Table would also be useful.

even in the absence of a repair command, it still would be useful
to see the bank table status. I was particularly curious if my RVT-H
had it's bank fully zero'd or if only a few values were unexpected.

this only adds the command to the CLI, as displaying raw bytes in a
QT window didn't feel like the best way to print that information.
If you want it to be there a UI before merging I understand, and can try
playing around with QT to get something decent.