[qrvthtool] ConfigStore::notifyAll(): Fix a deadlock if the config file has an invalid language tag.
The language tag is initially loaded by ConfigStore, but when notifyAll()
is called, LanguageMenu realizes it's invalid and calls ConfigStore::set()
with an empty tag, which causes a deadlock due to d->mtxSignalMaps being
locked.

To fix this, populate a single vector of methods to invoke, then unlock
the mutex and invoke the methods.